Addded option to return tmb object inside estimation object
Added an extra option to the estimation method which adds the tmb object
to the returned estimation object. This can then be used to directly vary
the parameters and quickly extract information from the filtration. This
could be also achieved by the tsfilter method with a specification object
but is much slower. Use case is for the multivariate GARCH partitioned
hessian calculation.